Move the client outside
